ENG-9278 - Create Validation Service and move ADV service to own class file
