+++ Package.resolved
@@ -18,6 +18,42 @@
           "revision": "51df49f935af022ceafbee34855500a37e0f35d3",
           "version": "2.7.0"
         }
+      },
+      {
+        "package": "JSONSchema",
+        "repositoryURL": "https://github.com/kylef/JSONSchema.swift",
+        "state": {
+          "branch": null,
+          "revision": "d14de4b2d9205068c9db89c00d097ca43c897000",
+          "version": "0.6.0"
+        }
+      },
+      {
+        "package": "DSJSONSchemaValidation",
+        "repositoryURL": "https://github.com/dashpay/JSONSchemaValidation",
+        "state": {
+          "branch": null,
+          "revision": "3b752520a969288702e987f3c9bfb3601aa2d720",
+          "version": "2.0.7"
+        }
+      },
+      {
+        "package": "PathKit",
+        "repositoryURL": "https://github.com/kylef/PathKit.git",
+        "state": {
+          "branch": null,
+          "revision": "3bfd2737b700b9a36565a8c94f4ad2b050a5e574",
+          "version": "1.0.1"
+        }
+      },
+      {
+        "package": "Spectre",
+        "repositoryURL": "https://github.com/kylef/Spectre.git",
+        "state": {
+          "branch": null,
+          "revision": "26cc5e9ae0947092c7139ef7ba612e34646086c7",
+          "version": "0.10.1"
+        }
       }
     ]
   },
