+++ Package.swift
@@ -14,7 +14,9 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/Alamofire/Alamofire.git", from: "5.0.0"),
-        .package(name: "FingerprintPro", url: "https://github.com/fingerprintjs/fingerprintjs-pro-ios", from: "2.7.0")
+        .package(name: "FingerprintPro", url: "https://github.com/fingerprintjs/fingerprintjs-pro-ios", from: "2.7.0"),
+        .package(name:"DSJSONSchemaValidation", url: "https://github.com/dashpay/JSONSchemaValidation", from: "2.0.7"),
+        .package(name: "JSONSchema", url:"https://github.com/kylef/JSONSchema.swift", from:"0.6.0")
 
     ],
     targets: [
@@ -28,6 +30,11 @@ let package = Package(
             resources: [
                 .process("PrivacyInfo.xcprivacy"),
             ]
+        ),
+        .testTarget(
+            name: "NeuroIDTests",
+            dependencies: ["NeuroID", "DSJSONSchemaValidation", "JSONSchema"],
+            path: "SDKTest"
         )
     ]
 )
